Add stream::extend()
