+++ src/collections/btree_set/from_stream.rs
@@ -17,7 +17,7 @@ impl<T: Ord> FromStream<T> for BTreeSet<T> {
             pin_utils::pin_mut!(stream);
 
             let mut out = BTreeSet::new();
-            stream::Extend::extend(&mut out, stream).await;
+            stream::extend(&mut out, stream).await;
             out
         })
     }
